bring Head to Head V2 tests up to component standards
